Component: Forbid connecting multiple pins to one signal
Do not allow to connect multiple symbol pins to one component signal.
This is done primary to reduce the complexity of the schematic editor.
Maybe we will revert this commit in future to re-allow connecting
multiple symbol pins to one component signal (this may be useful for
some components but requires to also make the schematic editor logic
more sophisticated).